Create helpers for running linting rules
